Add files via upload
 Election-Analysis

1.	OVERVIEW OF ELECTION AUDIT

•	Purpose Of This Election Analysis Audit

The election commission has requested some additional data to complete the audit:
•	The voter turnout for each county
•	The percentage of votes from each county out of the total count
•	The county with the highest turnout

2.	ELECTION-AUDIT RESULTS

•	Election Outcome
Total votes cast 
369,711 votes were cast in the election

B. Breakdown of the number of votes and the percentage of total votes for each county in the precinct.
County Votes:
The breakdown of the number of votes and percentage pf total votes for each county are as follows:
Jefferson:10.5% (38,855)
Denver:82.8% (306,055)
Arapahoe:6.7% (24,801)

C. County with the largest number of votes
Denver had the largest number of votes with 82.8% (306,055)

D. Breakdown of the number of votes and the percentage of the total votes each candidate received.
The breakdown of the number of votes and percentage pf total votes each candidate r	eceived are as follows:
Charles Casper Stockham: 23.0% (85,213)
Diana DeGette: 73.8% (272,892)
Raymon Anthony Doane: 3.1% (11,606)

E. Winner of the election
Diana DeGette won the election with 272,892 votes and 73.8% winning Percentage.


3. ELECTION-AUDIT SUMMARY

The election analysis shows that:
A.	There were 369,711 votes cast in the election using python/visual studio code for the analysis. Python scripts were created to perform calculations, iterations, logical operations, creations and additions of lists and keys to the dictionary. Repetition and decision statements were also used. 

B.	There were 369,711 total votes cast in the elections


C.	The county votes show that:
Jefferson scored 10.5% (38,855 votes)
Denver scored 82.8% (306,055 votes)
Arapahoe scored 6.7% (24,801 votes)

D.	Largest County Turnout:
 Denver scored 82.8% with 306,055 votes.

E.	Winner of the Election:
Winner: Diana DeGette
Winning Vote Count: 272,892
Winning Percentage: 73.8%
Recommendations to the election commissions
Visual studio codes (VS Code) can easily be used to analyze the outcome of the future elections for it is a free editor which helps programmers in error detection. It also supports multiple programming languages.

The codes used in this analysis are as follows:
•	County_vote = county_votes.get(county)
To retrieve the county vote count
•	County_vote_percentage = float(county_vote) / float(total_votes) * 100
To calculate the percentage of votes for the county ans so on.